gui_monitor/requirements.txt
awscrt==0.24.1
awsiotsdk==1.0.0.dev0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dotenv==1.1.0
shiboken6==6.9.1
